refactor: move StacksString and UrlString to net
