+++ src/core/mempool.rs
@@ -1109,8 +1109,8 @@ mod tests {
     use address::AddressHashMode;
     use burnchains::Address;
     use chainstate::burn::{BlockHeaderHash, VRFSeed};
-    use net::{Error as NetError, StacksMessageCodec};
-    use util::{hash::hex_bytes, hash::to_hex, hash::*, log, secp256k1::*, strings::StacksString};
+    use net::{Error as NetError, StacksMessageCodec, strings::StacksString};
+    use util::{hash::hex_bytes, hash::to_hex, hash::*, log, secp256k1::*};
     use vm::{
         database::HeadersDB,
         database::NULL_BURN_STATE_DB,
