+++ src/net/atlas/download.rs
@@ -11,7 +11,7 @@ use net::NeighborKey;
 use net::{GetAttachmentResponse, GetAttachmentsInvResponse};
 use net::{HttpRequestMetadata, HttpRequestType, HttpResponseType, PeerHost, Requestable};
 use util::hash::{Hash160, MerkleHashFunc};
-use util::strings;
+use net::strings;
 use util::{get_epoch_time_ms, get_epoch_time_secs};
 use vm::representations::UrlString;
 use vm::types::QualifiedContractIdentifier;
