+++ src/net/db.rs
@@ -59,7 +59,7 @@ use burnchains::PublicKey;
 
 use core::NETWORK_P2P_PORT;
 
-use util::strings::UrlString;
+use net::strings::UrlString;
 
 pub const PEERDB_VERSION: &'static str = "1";
